+++ jmx/jmx-command-client/src/main/java/uk/gov/justice/services/jmx/system/command/client/connection/ConnectorWrapper.java
@@ -1,7 +1,5 @@
 package uk.gov.justice.services.jmx.system.command.client.connection;
 
-import static java.util.Collections.emptyMap;
-
 import java.io.IOException;
 import java.util.Map;
 
@@ -11,8 +9,7 @@ import javax.management.remote.JMXServiceURL;
 
 public class ConnectorWrapper {
 
-    public JMXConnector connect(final JMXServiceURL serviceURL) throws IOException {
-        final Map<String, Object> environment = emptyMap();
+    public JMXConnector connect(final JMXServiceURL serviceURL, final Map<String, Object> environment) throws IOException {
         return JMXConnectorFactory.connect(serviceURL, environment);
     }
 }